Move ChangeCard to its own component
