+++ frontend/src/components/Dashboard.tsx
@@ -17,10 +17,10 @@ import ErrorOverlay from "@/components/ErrorOverlay";
 import Container from "@/components/Container";
 import PageLoader from "@/components/PageLoader";
 import CurrencyPairSelection from "@/components/CurrencyPairSelection";
+import { ChangeCard } from "@/components/ChangeCard";
 
 import type { CurrencyCode, ExchangeRate } from "@/types";
 import { CURRENCIES } from "@/currencies";
-import color from "@/styles/color";
 import {
   getAvailableCurrencyPairs,
   getHistoricalExchangeRates,
@@ -57,114 +57,6 @@ const SingleRateCard = ({
             {exchangeRate.rate.toFixed(4)}
           </Text>
         </Flex>
-      </CardContent>
-    </Card>
-  );
-};
-
-interface ChangeCardProps {
-  fromIsoCode: CurrencyCode;
-  toIsoCode: CurrencyCode;
-  currentExchangeRate: ExchangeRate;
-  previousExchangeRate: ExchangeRate | undefined;
-}
-
-const getBackgroundColorShade = (
-  relativeChangePercent: number,
-): 100 | 200 | 300 | 400 | 500 | 600 | 700 => {
-  const shadeBounds = [100, 700];
-  const changePercentBounds = [0.01, 5];
-
-  const interpolatedShade = Math.abs(
-    ((shadeBounds[1] - shadeBounds[0]) /
-      (changePercentBounds[1] - changePercentBounds[0])) *
-      (Math.abs(relativeChangePercent) - changePercentBounds[0]) +
-      shadeBounds[0],
-  );
-
-  const clamped = Math.min(
-    shadeBounds[1],
-    Math.max(shadeBounds[0], Math.round(interpolatedShade / 100) * 100),
-  ) as 100 | 200 | 300 | 400 | 500 | 600 | 700;
-
-  return clamped;
-};
-
-const ChangeCard = ({
-  fromIsoCode,
-  currentExchangeRate,
-  previousExchangeRate,
-}: ChangeCardProps) => {
-  if (!previousExchangeRate)
-    return <Text variant="danger">Data unavailable</Text>;
-
-  const { rate: currentValue, date: currentDate } = currentExchangeRate;
-  const { rate: previousValue, date: previousDate } = previousExchangeRate;
-
-  const numDaysDifference = currentDate.difference(previousDate, {
-    period: "days",
-  });
-
-  let headline = "";
-  if (numDaysDifference < 7) {
-    headline = `${numDaysDifference} days earlier`;
-  } else if (numDaysDifference < 30) {
-    const numWeeksDifference = Math.round(numDaysDifference / 7);
-    if (numWeeksDifference === 1) headline = "1 week earlier";
-    else headline = `${numWeeksDifference} weeks earlier`;
-  } else if (numDaysDifference < 365) {
-    const numMonthsDifference = Math.round(numDaysDifference / 30);
-    if (numMonthsDifference === 1) headline = "1 month earlier";
-    else headline = `${numMonthsDifference} months earlier`;
-  } else {
-    const numYearsDifference = Math.round(numDaysDifference / 365);
-    if (numYearsDifference === 1) headline = "1 year earlier";
-    else headline = `${Math.round(numDaysDifference / 365)} years earlier`;
-  }
-
-  const relativeChangePercent: Big = currentValue
-    .minus(previousValue)
-    .div(previousValue)
-    .times(100);
-
-  const relativeChangeDisplay = relativeChangePercent.gt(0)
-    ? `+${relativeChangePercent.toFixed(2)}%`
-    : `${relativeChangePercent.toFixed(2)}%`;
-
-  let subtext = "";
-
-  if (relativeChangePercent.abs().lt(0.1)) {
-    subtext = "No material change";
-  } else if (relativeChangePercent.gt(0)) {
-    subtext = `${fromIsoCode} stronger by ${relativeChangeDisplay}`;
-  } else if (relativeChangePercent.lt(0)) {
-    subtext = `${fromIsoCode} weaker by ${relativeChangeDisplay}`;
-  } else {
-    subtext = "N/A";
-  }
-
-  const backgroundColor = relativeChangePercent.gte(0) ? "green" : "red";
-  const backgroundColorShade = getBackgroundColorShade(
-    relativeChangePercent.toNumber(),
-  );
-
-  const colorValue = color[backgroundColor];
-
-  return (
-    <Card
-      className="w-full border py-0 outline-6 -outline-offset-8"
-      style={{
-        outlineColor: colorValue[backgroundColorShade] ?? colorValue["500"],
-      }}
-    >
-      <CardContent className="align-center flex flex-col justify-between gap-2 p-6">
-        <Text variant="muted">vs. {headline}</Text>
-        <Heading family="sans" level="2" numeric>
-          {relativeChangeDisplay}
-        </Heading>
-        <Text size="xs" variant="muted">
-          {subtext}
-        </Text>
       </CardContent>
     </Card>
   );
